# hw/17/api/openapi.yaml
openapi: 3.0.3
info:
  title: ctrsvc
  version: 1.0.0
  description: Учебный сервис users + orders (контракт для consumer).

paths:
  /healthz:
    get:
      operationId: getHealth
      summary: Health check
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Health'

  /users:
    get:
      operationId: listUsers
      summary: List users
      responses:
        '200':
          description: User list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      operationId: createUser
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'

  /users/{id}:
    get:
      operationId: getUser
      summary: Get user by id
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          $ref: '#/components/responses/NotFound'

  /orders:
    post:
      operationId: createOrder
      summary: Create order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'

  /orders/{id}:
    get:
      operationId: getOrder
      summary: Get order by id
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
        minLength: 1

  responses:
    BadRequest:
      description: Invalid request
      content:
        text/plain:
          schema:
            type: string
    NotFound:
      description: Resource not found
      content:
        text/plain:
          schema:
            type: string

  schemas:
    Health:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status:
          type: string
          enum: [ok]

    User:
      type: object
      additionalProperties: false
      required: [id, name, email]
      properties:
        id:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        email:
          type: string
          minLength: 1

    CreateUserRequest:
      type: object
      additionalProperties: false
      required: [name, email]
      properties:
        name:
          type: string
        email:
          type: string

    OrderItem:
      type: object
      additionalProperties: false
      required: [sku, qty]
      properties:
        sku:
          type: string
          minLength: 1
        qty:
          type: integer
          minimum: 1

    Order:
      type: object
      additionalProperties: false
      required: [id, user_id, total, items]
      properties:
        id:
          type: string
          minLength: 1
        user_id:
          type: string
          minLength: 1
        total:
          type: integer
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/OrderItem'

    CreateOrderRequest:
      type: object
      additionalProperties: false
      required: [user_id, items]
      properties:
        user_id:
          type: string
          minLength: 1
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/OrderItem'